customizable Text (#440)
* customizable Text

* update snapshots
