+++ src/components/Text/Text.styles.ts
@@ -1,10 +1,17 @@
 /*! Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license. */
-import { TextProps } from './Text';
-import { TextSize } from './types';
+import { TextSize, TextColor } from './Text.types';
 import { memoizeFunction } from '@uifabric/utilities';
-import { mergeStyleSets } from '@uifabric/styling';
+import { mergeStyleSets, ITheme } from '@uifabric/styling';
 import { ellipsisStyle, fontWeights, textColors, verticalAligns } from '../../util/styles/fonts';
-import { getTheme } from '../../util/colors';
+
+export interface TextClassNameProps {
+  color?: TextColor;
+  size?: TextSize;
+  maxWidth?: string;
+  bold?: boolean;
+  uppercase?: boolean;
+  theme: ITheme;
+}
 
 const iconSizeForSize = {
   [TextSize.XXLARGE]: '2.4rem',
@@ -16,8 +23,8 @@ const iconSizeForSize = {
   [TextSize.XSMALL]: '1.0rem',
 };
 
-const getMemoizedClassNames = memoizeFunction((styleProps: TextProps, theme) => {
-  const { size, maxWidth, bold, uppercase, color } = styleProps;
+export const getClassNames = memoizeFunction((classNameProps: TextClassNameProps) => {
+  const { size, maxWidth, bold, uppercase, color, theme } = classNameProps;
   const font = size ? theme.fonts[size === TextSize.MEDIUM_SUB ? 'smallPlus' : size] : undefined;
 
   return mergeStyleSets({
@@ -48,8 +55,3 @@ const getMemoizedClassNames = memoizeFunction((styleProps: TextProps, theme) => 
     },
   });
 });
-
-export const getClassNames = (styleProps: TextProps & { contextTextSize?: TextSize }) => {
-  const theme = getTheme();
-  return getMemoizedClassNames(styleProps, theme);
-};
